Refactor namespaces and update services
Refactored multiple classes to move them from `CompanionAPI.Models` to `CompanionAPI.Entities` namespace. Updated `UserOnboardingRequest` to change `Height` and `Weight` data types from `double` to `int`. Added new using directive in `DependencyInjection.cs` for `CompanionAPI.Services.GoalService`. Implemented `GetUserByEmailAsync` in `UserRepository`. Modified `CalcGoal` method in `GoalService` and `IGoalService` to accept additional parameters. Added error handling with new `GoalErrors` and `UserErrors` classes.

+++ CompanionAPI/Contracts/UserOnboardingContracts/UserOnboardingRequest.cs
@@ -5,6 +5,6 @@ public record UserOnboardingRequest(
     string Email,
     string Gender,
     int Age,
-    double Height,
-    double Weight
+    int Height,
+    int Weight
 );
